driver-geominimal-experiments/requirements.txt
git+https://gitlab+deploy-token-75:hAoek7ygRC_nNkKN83dX@gitlab.city-srv.ru/public_city/citymobil_python_clickhouse_wrapper.git@0.2.4#egg=citymobil_python_clickhouse_wrapper
git+https://gitlab+deploy-token-83:UVXbZBQUGhLdTgxsxtxy@gitlab.city-srv.ru/public_city/citymobil-python-mysql-wrapper.git@0.2.0#egg=citymobil_python_mysql_wrapper
pyexasol==0.14.3
python-dotenv==0.15.0
cffi==1.14.5
cryptography==3.4.5
newrelic==6.0.1.155
numpy==1.20.1
pandas==1.2.2
patsy==0.5.1
pyasn1==0.4.8
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
rsa==4.7
scipy==1.6.0
six==1.15.0
statsmodels==0.12.2
websocket-client==0.57.0
Shapely==1.7.1
pyproj==2.6.1.post1
aiohttp
pandasql
